refactor: Move the common logic to asset utils
